backend/pyproject.toml
[tool.poetry]
name = "booking(fastapi)"
version = "0.1.0"
description = "API for use booking service"
authors = ["Ilya Kibirev <kibirev.i@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.94.0"
uvicorn = "^0.21.0"
kafka-python = "^2.0.2"
pymongo = "^4.3.3"
motor = "^3.1.1"
mongogrant = "^0.3.3"
black = "^23.1.0"
isort = "^5.12.0"
python-decouple = "^3.8"
pydantic = {extras = ["dotenv"], version = "^1.10.6"}
loguru = "^0.6.0"
mypy = "0.981"
pre-commit = "^3.2.0"
pytz = "^2022.7.1"
python-dateutil = "^2.8.2"
orjson = "^3.8.8"
types-python-dateutil = "^2.8.19.10"
types-pytz = "^2022.7.1.2"


[tool.black]
color=true
exclude = ["./git", "./venv"]
include = '\.pyi?$'
line-length = 119

[tool.isort]
color_output = true
combine_as_imports = true
ensure_newline_before_comments = true
force_alphabetical_sort_within_sections = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
lines_between_sections = 1
multi_line_output = 3
profile = "black"
skip_gitignore = true
skip_glob = ["src/repository/migrations/**",]
src_paths = ["src/", "tests/",]
use_parentheses = true

[tool.mypy]
check_untyped_defs = true
color_output = true
error_summary = true
install_types = true
python_version = "3.11"
pretty = true
show_error_codes = true
strict_optional = true
warn_no_return = true
warn_return_any = false
ignore_missing_imports = true
exclude = "(src/config|logs|)/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
